Resolve conflicts and refactor order files
