+++ src/main/java/com/bootcamp/be_java_hisp_w29_g07/constants/Messages.java
@@ -9,6 +9,6 @@ public class Messages {
     public static final String USER_ALREADY_FOLLOW_SELLER = "The seller is already followed";
     public static final String USER_HAS_UNFOLLOWED_USER = "The user with id: %d has unfollowed user with id: %d.";
     public static final String USER_IS_NOT_FOLLOWING_USER = "The user with id: %d is not following the user with id: %d";
-    public static final String USER_HAS_NOT_FOLLOWED_MSG = "User %d has not followings";
+    public static final String USER_HAS_NOT_FOLLOWED_MSG = "User %d has no followings";
     public static final String USER_HAS_NOT_POSTS_MSG = "Followed by the user %d have no posts";
 }
